Assign default file extension based on audio format (#615)
I am using this library to record ogg/opus files, however, the default file name was always "sound.mp4" or "sound.m4a", regardless of the audio format.

To fix this, I tried passing in a custom path to startRecord, with the correct file extension, however a native extension is required to get the proper path to the cache directory on Android.

Since a native modification is required anyway, I decided to modify this module to just pick a reasonable default file extension for the giving audio format.

I also added the proper OGG / OPUS types for Android to index.ts. These types are supported as of apiVersion 29.