Fix race condition that was causing a crash in time observer (#628)
I was only able to reproduce this crash once, but I took a screenshot:

<img width="1182" alt="Screenshot 2024-09-15 at 5 01 46 PM"
src="https://github.com/user-attachments/assets/eddec4d4-2cbc-48da-89af-805ee49127cd">

The root cause is that `audioPlayer` is set to nil on the react-native
thread, not the main thread. So, even though we change for nil on line
338, there's no guarantee it's still not nil on line 340.

This PR will wrap the `audioPlayer = nil` mutation in
`DispatchQueue.main.async`.

I also changed the time observer to keep a weak reference to self,
because the apple docs state this is important to prevent a retain
cycle:
https://developer.apple.com/documentation/avfoundation/avplayer/1385829-addperiodictimeobserverforinterv?language=swift