fix: make `AudioRecorderPlayerImpl` internal to prevent Swift build errors (#680)
The actual implementation of this should probably be `internal`.
If it is `public`, Swift (or CocoaPods) is a bit stupid; it will try to
expose this to C++. If it does that, then it might cause cyclic
references or unknown symbols.

Unless you have a reason to make this `public`, we can keep it
`internal`.